[proliferate] split MCP materialization service
